Isolate module for Babel compiler
