+++ test/vm.spec.js
@@ -55,45 +55,5 @@ describe('vm.js', () => {
                     expect(alerts).to.be.an('array');
                 });
         });
-
-        it('should compile ES6 features using Babel compiler', () => {
-            const source = `
-                class Programmer {
-                    setName(name) { this.name = name; }
-                    getName() { return this.name }
-                    greet() { return \`Hello, I'm \${this.name}\`; }
-                }
-
-                export default { extension: () => {
-                    const hello = new Programmer();
-                    hello.setName('Marcelo');
-                    return hello.greet();
-                } };
-            `;
-
-            return runAndGetAlerts({ name: 'test-babel-comnpiler', source }, {})
-                .then(result => {
-                    expect(result).to.equals('Hello, I\'m Marcelo');
-                });
-        });
-
-        it('should compile JSX syntax down to HTML', () => {
-            const source = `
-                const annoy = 'aborrecer';
-                export default { extension: () => {
-                    return (
-                        <div className="component">
-                            Você fala demais, acabei de me <b>{ annoy }</b>
-                            <br hue="land" />
-                        </div>
-                    );
-                } }
-            `;
-
-            return runAndGetAlerts({ name: 'test-jsx-compiler', source }, {})
-                .then(result => {
-                    console.log(result);
-                });
-        });
     })
 });
